shift, collisions fixed ..
git-svn-id: http://undergroundlab.cz/svn/dark_mir@80 e2d330a7-c6b2-4f77-9a9a-126ff9ba3f6e

+++ protokol.c
@@ -335,7 +335,7 @@ int Get_object_states(){
 //==============================================================================	
 
 	Uint8 *tp = r->data;
-	Uint8 id = 0;
+	Uint32 id = 0;
 
 
   if((*tp != P_OBJECT_STATES) &&
